Use composable to factorize code between PlanNode and PlanNodeDetail
